requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
dominate==2.6.0
Flask==2.1.2
Flask-Bootstrap==3.3.7.1
Flask-Reuploaded==1.2.0
Flask-WTF==1.0.1
google-api-core==2.7.3
google-auth==2.6.6
google-cloud-core==2.3.0
google-cloud-firestore==2.4.0
googleapis-common-protos==1.56.1
grpcio==1.46.1
grpcio-status==1.46.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
proto-plus==1.20.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
